avformat/factory.c, jackrack/jack_rack.c, jackrack/plugin_settings.c, vmfx/filter_chroma.c, plus/transition_affine.c, westley/producer_westley.c, westley/consumer_westley.c, kdenlive/filter_wave.c, vorbis/producer_vorbis.c, gtk2/producer_pango.c, gtk2/pixops.c, miracle_server.c, miracle_unit.c: cleanup a whole bunch of compiler warnings
git-svn-id: https://mlt.svn.sourceforge.net/svnroot/mlt/trunk/mlt@1035 d19143bc-622f-0410-bfdd-b5b2a6649095

+++ src/miracle/miracle_unit.c
@@ -593,7 +593,7 @@ void miracle_unit_change_position( miracle_unit unit, int clip, int32_t position
 	else if ( clip >= mlt_playlist_count( playlist ) )
 	{
 		clip = mlt_playlist_count( playlist ) - 1;
-		position = LONG_MAX;
+		position = INT_MAX;
 	}
 
 	if ( mlt_playlist_get_clip_info( playlist, &info, clip ) == 0 )
